bonus/api/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.22
scikit-learn>=1.3.0
